move tracking event
