+++ src/app/Scenes/Consignments/Screens/SubmitArtworkOverview/utils/TrackingEvent.ts
